# expression: implement vectorized evaluation for builtinSysDateWithoutFspSig